requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyNaCl==1.5.0
schwab-py==1.5.1
